[refactor] 게시글 작성하기 페이지 클라이언트 분리
